refactor: migrate queue job functionality to a dedicated utility module
