+++ src/Signal/Core/session_builder.ts
@@ -1,12 +1,12 @@
 // @ts-nocheck
 'use strict'
 
+import queueJob from '../../Utils/queue-job'
 import BaseKeyType from './base_key_type'
 import ChainType from './chain_type'
 import * as crypto from './crypto'
 import * as curve from './curve'
 import * as errors from './errors'
-import queueJob from './queue_job'
 import SessionRecord from './session_record'
 
 class SessionBuilder {
