+++ src/Utils/queue-job.ts
@@ -1,3 +1,5 @@
+// Type-safe async job queue utility for serializing operations per bucket key.
+
 interface QueueJob<T> {
 	awaitable: () => Promise<T>
 	resolve: (value: T | PromiseLike<T>) => void
@@ -36,12 +38,12 @@ async function _asyncQueueExecutor(queue: Array<QueueJob<any>>, cleanup: () => v
 	cleanup()
 }
 
+/**
+ * Serializes async operations for a given bucket key.
+ * @param bucket string or number key to identify the queue
+ * @param awaitable async function to execute
+ */
 export default function queueJob<T>(bucket: string | number, awaitable: () => Promise<T>): Promise<T> {
-	// Skip name assignment since it's readonly in strict mode
-	if (typeof bucket !== 'string') {
-		console.warn('Unhandled bucket type (for naming):', typeof bucket, bucket)
-	}
-
 	let inactive = false
 	if (!_queueAsyncBuckets.has(bucket)) {
 		_queueAsyncBuckets.set(bucket, [])
